refactor geometry submodule
