+++ src/nerte/geometry/__init__.py
@@ -1,0 +1,3 @@
+"""Module for geometries."""
+
+from .base import Geometry, intersection_ray_depth, StandardGeometry
