+++ src/nerte/render/image_color_renderer_unittest.py
@@ -18,7 +18,7 @@ from nerte.values.color import Colors
 from nerte.world.object import Object
 from nerte.world.camera import Camera
 from nerte.world.scene import Scene
-from nerte.geometry.geometry import StandardGeometry
+from nerte.geometry import StandardGeometry
 from nerte.render.projection import ProjectionMode
 from nerte.render.image_color_renderer import ImageColorRenderer
