+++ src/nerte/render/image_filter_renderer.py
@@ -15,7 +15,7 @@ from nerte.values.intersection_info import IntersectionInfo, IntersectionInfos
 from nerte.world.camera import Camera
 from nerte.world.object import Object
 from nerte.world.scene import Scene
-from nerte.geometry.geometry import Geometry
+from nerte.geometry import Geometry
 from nerte.render.image_renderer import ImageRenderer
 from nerte.render.projection import ProjectionMode
 from nerte.util.generic_matrix import GenericMatrix
